IC-47 fix for capital_limit parameter on resource_allocator; rework for binance_wrapper_cli; fix for _execute_cancel in case of EState.ENTER_EXP; update TestEnterExpire for live-trade; logging improved for trades and df_balance

+++ src/Ikarus/utils.py
@@ -141,7 +141,7 @@ def round_to_period(current_time, period_min, direction='floor', offset=0):
 
 
 def setup_logger(logger, log_config):
-    if log_config.get('clear',False):
+    if log_config.get('clear',False) and os.path.isfile(log_config['file']) :
         os.remove(log_config['file'])
 
     logger = logging.getLogger('app')
